Feat/map upload

+++ backend/src/main/java/com/webgis/map/finalmap/FinalMapRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface FinalMapRepository extends JpaRepository<FinalMap, Integer> {
+public interface FinalMapRepository extends JpaRepository<FinalMap, Long> {
     Optional<FinalMap> findByTitle(String title);
     Optional<FinalMap> findById(long id);
 }
